restvoteapp-master/src/main/java/restvoteapp/service/UserService.java
package restvoteapp.service;

import lombok.AllArgsConstructor;
import org.springframework.context.annotation.Scope;
import org.springframework.context.annotation.ScopedProxyMode;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import restvoteapp.AuthorizedUser;
import restvoteapp.model.User;
import restvoteapp.repository.UserRepository;
import restvoteapp.util.exception.NotFoundException;

@Service("userService")
@Scope(proxyMode = ScopedProxyMode.TARGET_CLASS)
@AllArgsConstructor
public class UserService implements UserDetailsService {

    private final UserRepository userRepository;

    @Override
    public AuthorizedUser loadUserByUsername(String email) throws UsernameNotFoundException {
        User user = userRepository.getByEmail(email.toLowerCase());
        if (user == null) {
            throw new UsernameNotFoundException("User " + email + " is not found");
        }
        return new AuthorizedUser(user);
    }

    public User getById(int id) {
        return userRepository.findById(id).orElseThrow(() -> new NotFoundException("Not found user with id " + id));
    }
}
